% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matches.R
\name{matches}
\alias{matches}
\title{matches}
\usage{
matches(
  match_id,
  api_key = NULL,
  base_url = "https://api.opendota.com/api/",
  path = "matches/"
)
}
\arguments{
\item{match_id}{match_id}

\item{api_key}{optional API key}

\item{base_url}{base_url}

\item{path}{path}
}
\value{
a list
}
\description{
returns a list with various data for the given match_id
}
