% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opendota.api-package.R
\docType{package}
\name{opendota.api-package}
\alias{opendota.api}
\alias{opendota.api-package}
\title{opendota.api: Interface for the opendota.com API}
\description{
Interface for the opendota.com API.
}
\keyword{internal}
