% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsed_matches.R
\name{parsed_matches}
\alias{parsed_matches}
\title{parsed_matches}
\usage{
parsed_matches(
  base_url = "https://api.opendota.com/api/",
  path = "parsedMatches/",
  less_than_match_id = NULL
)
}
\arguments{
\item{base_url}{base_url}

\item{path}{path}

\item{less_than_match_id}{less_than_match_id}
}
\value{
an integer vector
}
\description{
returns available matches with match id smaller than less_than_match_id
}
