% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/players_by_rank.R
\name{players_by_rank}
\alias{players_by_rank}
\title{players_by_rank}
\usage{
players_by_rank(
  base_url = "https://api.opendota.com/api/",
  path = "playersByRank/"
)
}
\value{
A data.table
}
\description{
players_by_rank
}
\examples{
players_by_rank()
}
